+++ src/content/Share.css
@@ -8,10 +8,6 @@
   justify-content: center;
 }
 
-.share-twist .twist-body {
-  display: block;
-}
-
 .r2c-twist .twist .share-twist .tweet-icon {
   display: hidden;
 }
